+++ lib/game.rb
@@ -1,58 +1,3 @@
-# print boars for the user to see
-class UserDisplay
-  def self.print_board(grid)
-    p '********************'
-    grid.length.times { |n| p grid[n] }
-    p '********************'
-  end
-
-  def wich_turn(player_1_name, player_2_name)
-    player_turn_name = player_1_name if $current_turn.even?
-    mark = 'X' if $current_turn.even?
-    player_turn_name = player_2_name if $current_turn.odd?
-    mark = 'O' if $current_turn.odd?
-    p "It's #{player_turn_name}'s turn' ('#{mark}'), make your move:"
-  end
-end
-
-# Get the player movement and procces it to se if it's valid
-class PlayerMoveInput
-  @@valid_moves = %w[1 2 3 4 5 6 7 8 9]
-  def validate_move(player_mark, modified_grid, player_1_name, player_2_name)
-    valid_move = false
-    @@valid_moves.length.times do |n|
-      if player_mark == @@valid_moves[n]
-        valid_move = true
-        @@valid_moves.delete(player_mark)
-        CheckGameOn.new.draw(@@valid_moves)
-        CheckGameOn.new.player_move_save(player_mark, player_1_name, player_2_name)
-        move_print(player_mark, modified_grid)
-        UserDisplay.new.wich_turn(player_1_name, player_2_name) if $game_on
-        return UserDisplay.print_board(modified_grid)
-      end
-    end
-    unless valid_move
-      p "That's not a valid move, try again" if $game_on
-      UserDisplay.new.wich_turn(player_1_name, player_2_name) if $game_on
-      UserDisplay.print_board(modified_grid) if $game_on
-    end
-  end
-
-  def move_print(player_mark_print, modified_grid_print)
-    player_turn_print = 'X' if $current_turn.even?
-    player_turn_print = 'O' if $current_turn.odd?
-    modified_grid_print.each do |sub_array|
-      sub_array.length.times do |n|
-        if player_mark_print == sub_array[n]
-          sub_array[n] = player_turn_print
-          $current_turn += 1
-          return modified_grid_print
-        end
-      end
-    end
-  end
-end
-
 # check to see if the game should be on or off
 class CheckGameOn
   @@winning_cases = [%w[1 2 3], %w[4 5 6], %w[7 8 9], %w[1 4 7], %w[2 5 8], %w[3 6 9], %w[1 5 9], %w[3 5 7]]
